@@ -1,6 +1,6 @@
 {
   "name": "docs",
-  "version": "4.0.0-alpha.37",
+  "version": "4.0.0",
   "private": true,
   "author": "MUI Team",
   "license": "MIT",
@@ -9,7 +9,7 @@
     "build:clean": "rimraf .next && yarn build",
     "build-sw": "node ./scripts/buildServiceWorker.js",
     "dev": "rimraf ./node_modules/.cache/babel-loader && next dev --port 3001",
-    "deploy": "git push upstream master:release",
+    "deploy": "git push upstream next:docs-v5",
     "export": "rimraf docs/export && next export --threads=3 -o export && yarn build-sw",
     "icons": "rimraf public/static/icons/* && node ./scripts/buildIcons.js",
     "typescript": "tsc -p tsconfig.json",
@@ -17,17 +17,17 @@
     "typescript:transpile:dev": "node scripts/formattedTSDemos --watch"
   },
   "dependencies": {
-    "@babel/core": "^7.15.0",
+    "@babel/core": "^7.15.5",
     "@babel/plugin-transform-object-assign": "^7.14.5",
-    "@babel/runtime-corejs2": "^7.15.3",
-    "@material-ui/core": "^4.9.12",
-    "@material-ui/docs": "^4.0.0-beta.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.56",
-    "@material-ui/monorepo": "https://github.com/mui-org/material-ui.git#master",
+    "@babel/runtime-corejs2": "^7.15.4",
+    "@emotion/cache": "^11.4.0",
+    "@emotion/react": "^11.4.1",
+    "@emotion/server": "^11.4.0",
+    "@emotion/styled": "^11.3.0",
     "@mui/icons-material": "^5.0.0-rc.1",
     "@mui/material": "^5.0.0-rc.1",
     "@mui/styles": "^5.0.0-rc.1",
+    "@mui/utils": "^5.0.0-rc.1",
     "@trendmicro/react-interpolate": "^0.5.5",
     "@types/react-dom": "^17.0.9",
     "@types/react-router-dom": "^5.1.8",
@@ -56,8 +56,8 @@
     "lodash": "^4.17.21",
     "lz-string": "^1.4.4",
     "markdown-to-jsx": "^7.1.3",
-    "marked": "^2.1.3",
-    "next": "^9.5.0",
+    "marked": "^3.0.3",
+    "next": "^11.0.1",
     "nprogress": "^0.2.0",
     "postcss": "^8.3.6",
     "prismjs": "^1.24.1",
@@ -67,14 +67,15 @@
     "react-docgen": "^5.4.0",
     "react-dom": "^17.0.2",
     "react-is": "^17.0.2",
-    "react-redux": "^7.2.4",
     "react-router": "^5.2.1",
     "react-router-dom": "^5.2.1",
     "react-text-mask": "^5.4.3",
     "recast": "^0.20.5",
-    "redux": "^4.1.1",
-    "redux-logger": "^3.0.6",
     "rimraf": "^3.0.2",
+    "styled-components": "^5.0.0",
+    "stylis": "^4.0.3",
+    "stylis-plugin-rtl": "^2.1.0",
+    "stylis-plugin-rtl-sc": "npm:stylis-plugin-rtl@^1.1.0",
     "webfontloader": "^1.6.28",
     "webpack-bundle-analyzer": "^4.4.2"
   },
