@@ -1,6 +1,6 @@
 import * as React from 'react';
-import Select from '@material-ui/core/Select';
-import MenuItem from '@material-ui/core/MenuItem';
+import Select from '@mui/material/Select';
+import MenuItem from '@mui/material/MenuItem';
 import { DataGrid, GridRenderEditCellParams } from '@mui/x-data-grid';
 
 const SelectEditInputCell = (props: GridRenderEditCellParams) => {
