@@ -1,4 +1,4 @@
-import { ponyfillGlobal } from '@material-ui/utils';
+import { ponyfillGlobal } from '@mui/utils';
 
 // eslint-disable-next-line no-underscore-dangle
 ponyfillGlobal.__MUI_RELEASE_INFO__ = 'MTU5NjMxOTIwMDAwMA=='; // 2020-08-02
