@@ -437,11 +437,10 @@ export function useGridEditRows(
   );
 
   const handleCellEditStop = React.useCallback(
-    (params: GridCellParams, event?: React.SyntheticEvent) => {
+    (params: GridCellParams, event: React.SyntheticEvent | {}) => {
       setCellMode(params.id, params.field, GridCellModes.View);
 
-      // When dispatched by the document, the event is not passed
-      if (!event || !isKeyboardEvent(event)) {
+      if (!isKeyboardEvent(event)) {
         return;
       }
 
