@@ -1,3 +1,7 @@
-export type MuiEvent<E> = E & {
+import * as React from 'react';
+
+type BaseEvent = React.SyntheticEvent | DocumentEventMap[keyof DocumentEventMap] | {};
+
+export type MuiEvent<E extends BaseEvent = BaseEvent> = E & {
   defaultMuiPrevented?: boolean;
 };
